requirements.txt
ale-py==0.7.4
einops==0.3.2
gymnasium[accept-rom-license]==0.27.0
hydra-core==1.1.1
minigrid==2.1.1
numpy>=1.18.0
opencv-python
protobuf==3.20.*
psutil==5.8.0
pygame==2.1.2
torch==1.13.1
torchvision==0.14.1
Pillow==9.4.0
tqdm==4.62.3
wandb==0.12.7
